% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPrinStratTTE-package.R
\docType{package}
\name{BPrinStratTTE-package}
\alias{BPrinStratTTE-package}
\alias{BPrinStratTTE}
\title{The 'BPrinStratTTE' package.}
\description{
Bayesian models to estimate causal effects of biological
treatments on time-to-event endpoints in clinical trials with principal
strata defined by the occurrence of antidrug antibodies.
The methodology is based on Frangakis and Rubin (2002)
\url{doi:10.1111/j.0006-341x.2002.00021.x} and Imbens and Rubin (1997)
\url{doi:10.1214/aos/1034276631}, and intended to be applied to a
specific time-to-event setting.#'
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Boehringer-Ingelheim/BPrinStratTTE}
  \item Report bugs at \url{https://github.com/Boehringer-Ingelheim/BPrinStratTTE/issues}
}

}
\author{
\strong{Maintainer}: Christian Stock \email{christian.stock@boehringer-ingelheim.com} (\href{https://orcid.org/0000-0002-3493-3234}{ORCID})

Authors:
\itemize{
  \item Emma Torrini
}

Other contributors:
\itemize{
  \item Boehringer Ingelheim Pharma GmbH & Co. KG [copyright holder, funder]
}

}
