% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mult_exp_nocovar.R
\name{fit_mult_exp_nocovar}
\alias{fit_mult_exp_nocovar}
\title{Fit multiple models without covariates}
\usage{
fit_mult_exp_nocovar(list_dfs, fit_function, model, m_params, seed = 23)
}
\arguments{
\item{list_dfs}{...}

\item{fit_function}{...}

\item{model}{...}

\item{m_params}{...}
}
\value{
...
}
\description{
Fit multiple models without covariates
}
\examples{
print("...")

}
