% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_single_exp_nocovar.R
\name{fit_single_exp_nocovar}
\alias{fit_single_exp_nocovar}
\title{Fit Bayesian principal stratification model for an exponentially distributed endpoint without consideration of predictors of the intercurrent event}
\usage{
fit_single_exp_nocovar(data, model, params)
}
\arguments{
\item{data}{...}

\item{model}{...}

\item{params}{...}
}
\value{
...
}
\description{
Fit Bayesian principal stratification model for an exponentially distributed endpoint without consideration of predictors of the intercurrent event
}
\examples{
print("...")

}
