% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_exp_nocovar.R
\name{run_sim_exp_nocovar}
\alias{run_sim_exp_nocovar}
\title{Run simulation of models without covariates}
\usage{
run_sim_exp_nocovar(n_iter, d_params, fit_function, model, m_params)
}
\arguments{
\item{n_iter}{...}

\item{d_params}{...}

\item{fit_function}{...}

\item{model}{...}

\item{m_params}{...}
}
\value{
...
}
\description{
Run simulation of models without covariates
}
\examples{
print("...")

}
