% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_mult_trials_exp_nocovar.R
\name{sim_dat_mult_trials_exp_nocovar}
\alias{sim_dat_mult_trials_exp_nocovar}
\title{Simulate data from multiple 2-arm trials without predictors of intercurrent events}
\usage{
sim_dat_mult_trials_exp_nocovar(n_iter, params)
}
\arguments{
\item{n_iter}{...}

\item{params}{...}
}
\value{
...
}
\description{
Simulate data from multiple 2-arm trials without predictors of intercurrent events
}
\examples{
print("...")

}
