% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_one_trial_exp_covar.R
\name{sim_dat_one_trial_exp_covar}
\alias{sim_dat_one_trial_exp_covar}
\title{Simulate data from a single 2-arm trial with a predictor of the intercurrent events}
\usage{
sim_dat_one_trial_exp_covar(
  n,
  nt,
  prob_X1,
  prob_ice_X1,
  prob_ice_X0,
  fu_max,
  T0T_rate,
  T0N_rate,
  T1T_rate,
  T1N_rate
)
}
\arguments{
\item{n}{...}

\item{nt}{...}

\item{prob_X1}{...}

\item{prob_ice_X1}{...}

\item{prob_ice_X0}{...}

\item{fu_max}{...}

\item{T0T_rate}{...}

\item{T0N_rate}{...}

\item{T1T_rate}{...}

\item{T1N_rate}{...}
}
\value{
...
}
\description{
Simulate data from a single 2-arm trial with a predictor of the intercurrent events
}
\examples{
d_params_covar <- list(
  n = 1000,        
  nt = 500,       
  prob_X1 = 0.4, 
  prob_ice_X1 = 0.5, 
  prob_ice_X0 = 0.2,
  fu_max = 48*7,       
  T0T_rate = 0.2,     
  T0N_rate = 0.2,     
  T1T_rate = 0.15,     
  T1N_rate = 0.1
 )
dat_single_trial <- sim_dat_one_trial_exp_covar(
  n = d_params_covar[["n"]], 
  nt = d_params_covar[["nt"]],
  prob_X1 = d_params_covar[["prob_X1"]],
  prob_ice_X1 = d_params_covar[["prob_ice_X1"]],
  prob_ice_X0 = d_params_covar[["prob_ice_X0"]],
  fu_max = d_params_covar[["fu_max"]],  
  T0T_rate = d_params_covar[["T0T_rate"]],
  T0N_rate = d_params_covar[["T0N_rate"]],
  T1T_rate = d_params_covar[["T1T_rate"]],
  T1N_rate = d_params_covar[["T1N_rate"]] 
)
dim(dat_single_trial)
head(dat_single_trial)
                  
}
