% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_one_trial_exp_nocovar.R
\name{sim_dat_one_trial_exp_nocovar}
\alias{sim_dat_one_trial_exp_nocovar}
\title{Simulate data from a single 2-arm trial without predictors of intercurrent events}
\usage{
sim_dat_one_trial_exp_nocovar(
  n,
  nt,
  prob_ice,
  fu_max,
  T0T_rate,
  T0N_rate,
  T1T_rate,
  T1N_rate
)
}
\arguments{
\item{n}{Number of subjects in the trial}

\item{nt}{Number of treated subjects}

\item{prob_ice}{Probability of an intercurrent event}

\item{fu_max}{Maximum follow-up time in days (administrative censoring is assumed afterwards)}

\item{T0T_rate}{Monthly event rate in control subjects that would develop the intercurrent event if given treatment}

\item{T0N_rate}{Monthly event rate in control subjects that never develop the intercurrent event}

\item{T1T_rate}{Monthly event rate in treated subjects that develop the intercurrent event}

\item{T1N_rate}{Monthly event rate in treated subjects that never develop the intercurrent event}
}
\value{
A \code{tibble()}containing an analysis dataset
}
\description{
Simulate data from a single 2-arm trial without predictors of intercurrent events
}
\examples{
print("...")
}
