% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPrinStratTTE-package.R
\docType{package}
\name{BPrinStratTTE-package}
\alias{BPrinStratTTE}
\alias{BPrinStratTTE-package}
\title{BPrinStratTTE: Bayesian principal stratification for time-to-event endpoints}
\description{
Functions to fit Bayesian principal stratification models for exponentially distributed time-to-event endpoints.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chstock/BPrinStratTTE}
  \item Report bugs at \url{https://github.com/chstock/BPrinStratTTE/issues}
}

}
\author{
\strong{Maintainer}: Christian Stock \email{christian.stock@boehringer-ingelheim.com} (\href{https://orcid.org/0000-0002-3493-3234}{ORCID})

Authors:
\itemize{
  \item Emma Torrini
}

Other contributors:
\itemize{
  \item Boehringer Ingelheim Pharma GmbH & Co. KG [copyright holder, funder]
}

}
\keyword{internal}
