% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs_exp_nocovar.R
\name{ocs_exp_nocovar}
\alias{ocs_exp_nocovar}
\title{Determine operating characteristics of multiple models}
\usage{
ocs_exp_nocovar(multiple_fits, d_params, m_params)
}
\arguments{
\item{multiple_fits}{...}

\item{d_params}{...}

\item{m_params}{...}
}
\value{
...
}
\description{
Determine operating characteristics of multiple models
}
\examples{
print("...")
}
