% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_one_trial_exp_nocovar.R
\name{sim_dat_one_trial_exp_nocovar}
\alias{sim_dat_one_trial_exp_nocovar}
\title{Simulate data from a single 2-arm trial without predictors of intercurrent events}
\usage{
sim_dat_one_trial_exp_nocovar(
  n,
  nt,
  prob_ice,
  fu_max,
  T0T_rate,
  T0N_rate,
  T1T_rate,
  T1N_rate
)
}
\arguments{
\item{n}{Number of subjects in the trial}

\item{nt}{Number of treated subjects}

\item{prob_ice}{Probability of an intercurrent event}

\item{fu_max}{Maximum follow-up time in days (administrative censoring is assumed afterwards)}

\item{T0T_rate}{Monthly event rate in control subjects that would develop the intercurrent event if given treatment}

\item{T0N_rate}{Monthly event rate in control subjects that never develop the intercurrent event}

\item{T1T_rate}{Monthly event rate in treated subjects that develop the intercurrent event}

\item{T1N_rate}{Monthly event rate in treated subjects that never develop the intercurrent event}
}
\value{
A \code{tibble()}containing an analysis dataset
}
\description{
Simulate data from a single 2-arm trial without predictors of intercurrent events
}
\examples{
d_params_nocovar <- list(
  n = 500L,
  nt = 250L,
  prob_ice = 0.5,
  fu_max = 336L,
  T0T_rate = 0.2,
  T0N_rate = 0.2,
  T1T_rate = 0.15,
  T1N_rate = 0.1
)
dat <- sim_dat_one_trial_exp_nocovar(
  n = d_params_nocovar[["n"]], 
  nt = d_params_nocovar[["nt"]],
  prob_ice = d_params_nocovar[["prob_ice"]],
  fu_max = d_params_nocovar[["fu_max"]],  
  T0T_rate = d_params_nocovar[["T0T_rate"]],
  T0N_rate = d_params_nocovar[["T0N_rate"]],
  T1T_rate = d_params_nocovar[["T1T_rate"]],
  T1N_rate = d_params_nocovar[["T1N_rate"]] 
)
dim(dat)
head(dat)

}
