% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_one_trial_exp_nocovar.R
\name{sim_dat_one_trial_exp_nocovar}
\alias{sim_dat_one_trial_exp_nocovar}
\title{Simulate data from a single two-arm trial with an exponentially distributed time-to-event endpoint and no predictor of the intercurrent event}
\usage{
sim_dat_one_trial_exp_nocovar(
  n,
  nt,
  prob_ice,
  fu_max,
  prop_cens = 0,
  T0T_rate,
  T0N_rate,
  T1T_rate,
  T1N_rate
)
}
\arguments{
\item{n}{Positive integer value, number of subjects in the trial.}

\item{nt}{Positive integer value, number of treated subjects.}

\item{prob_ice}{Numeric value on the interval \eqn{(0,1)}, probability of the intercurrent event of interest (i.e. the event that determines the principal stratum membership).}

\item{fu_max}{Positive integer value, maximum follow-up time in days (administrative censoring assumed afterwards).}

\item{prop_cens}{Numeric value on the interval \eqn{[0,1)}, proportion of uniformly censored patients (default is 0).}

\item{T0T_rate}{Positive numeric value, monthly event rate in control subjects that would develop the intercurrent event if treated.}

\item{T0N_rate}{Positive numeric value, monthly event rate in control subjects that never develop the intercurrent event.}

\item{T1T_rate}{Positive numeric value, monthly event rate in treated subjects that develop the intercurrent event.}

\item{T1N_rate}{Positive numeric value, monthly event rate in treated subjects that never develop the intercurrent event.}
}
\value{
A \code{tibble()}containing the trial data for analysis.
}
\description{
Simulate data from a single two-arm trial with an exponentially distributed time-to-event endpoint and no predictor of the intercurrent event
}
\examples{
d_params_nocovar <- list(
  n = 500L,
  nt = 250L,
  prob_ice = 0.5,
  fu_max = 336L,
  prop_cens = 0.15,
  T0T_rate = 0.2,
  T0N_rate = 0.2,
  T1T_rate = 0.15,
  T1N_rate = 0.1
)
dat_single_trial <- sim_dat_one_trial_exp_nocovar(
  n = d_params_nocovar[["n"]], 
  nt = d_params_nocovar[["nt"]],
  prob_ice = d_params_nocovar[["prob_ice"]],
  fu_max = d_params_nocovar[["fu_max"]],
  prop_cens = d_params_nocovar[["prop_cens"]],  
  T0T_rate = d_params_nocovar[["T0T_rate"]],
  T0N_rate = d_params_nocovar[["T0N_rate"]],
  T1T_rate = d_params_nocovar[["T1T_rate"]],
  T1N_rate = d_params_nocovar[["T1N_rate"]] 
)
dim(dat_single_trial)
head(dat_single_trial)

}
\seealso{
\code{\link[=sim_dat_one_trial_exp_covar]{sim_dat_one_trial_exp_covar()}}
}
